+++ examples/redis/src/main.rs
@@ -1,7 +1,7 @@
 use anyhow::Context;
-use autumn::App;
-use autumn_redis::{redis::AsyncCommands, Redis, RedisPlugin};
-use autumn_web::{
+use spring::App;
+use spring_redis::{redis::AsyncCommands, Redis, RedisPlugin};
+use spring_web::{
     error::Result,
     extractor::{Component, Path},
     get,