+++ spring/src/lib.rs
@@ -1,0 +1,2 @@
+pub use spring_boot::app::App;
+pub use spring_macros::*;